+++ src/handlers/index.ts
@@ -1,5 +1,5 @@
 export { handleError } from "./error.js";
-export { normalizeHeadMethod, toHeadResponse } from "./head.js";
+export { toHeadResponse } from "./head.js";
 export { handleMethodNotAllowed } from "./methodNotAllowed.js";
 export { handleNotFound } from "./notFound.js";
 export { handleOptions } from "./options.js";