+++ src/routes/index.ts
@@ -1,3 +1,10 @@
 export { compilePattern, createRouteBuilder, wrapBuilderWithNamePrefix } from "./builder.js";
-export { findRoute, getAllowedMethods, hasMatchingPath, type RouteMatch } from "./find.js";
+export {
+	findRoute,
+	getAllowedMethods,
+	hasMatchingPath,
+	type RouteMatch,
+	type RouteResolution,
+	resolveRoute,
+} from "./find.js";
 export { type AddRouteFn, createGroupRouter } from "./group.js";